moved common code to common dir
