+++ zholobov.oleg/common/read_string.hpp
@@ -1,0 +1,10 @@
+#ifndef READ_STRING_HPP
+#define READ_STRING_HPP
+
+#include <istream>
+
+namespace zholobov {
+  char* read_string(std::istream& input);
+}
+
+#endif
